blogpostAutomation/func-blogpost-automation/requirements.txt
# Minimal requirements for blog automation
# Azure Functions
azure-functions==1.18.0

# HTTP Requests (for Perplexity API)
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Telegram Bot
pyTelegramBotAPI==4.14.0

# Git Operations (simpler alternative)
dulwich==0.21.7

# Date/Time handling
python-dateutil==2.8.2

# JSON handling (built-in json works too)
# Basic logging (built-in logging works)

# Retry logic
tenacity==8.2.3